// src/api/schema.d.ts
export interface components {
    schemas: {
        CreateMonsterRequest: {
            name: string;
            hp: number;
            attack: number;
            defense: number;
            speed: number;
            /** Format: uri */
            imageUrl: string;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}